refactor: split out kafka metrics and parser
